yazlabBackEnd/requirement.txt
asgiref==3.8.1
bcrypt==4.1.3
click==8.1.7
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
joblib==1.4.2
nltk==3.8.1
numpy==1.26.4
psycopg2==2.9.9
pybind11==2.12.0
PyJWT==2.8.0
rake-nltk==1.0.6
regex==2024.5.15
scikit-learn==1.4.2
scipy==1.13.0
sqlparse==0.5.0
summa==1.2.0
threadpoolctl==3.5.0
tqdm==4.66.4

